+++ README.md
@@ -19,6 +19,8 @@ We extend H3 to include the following concerns, some of which we tend to use acr
 portfolio of services and some of which are unique to H3:
 
 <dl>
+  <dt>CocinaBuilders</dt>
+  <dd>Build Cocina data structures from primitive data types</dd>
   <dt>Components</dt>
   <dd>Provide views for reusable UI elements and easier testing (from <a href="https://viewcomponent.org/">ViewComponent</a> gem)</dd>
   <dt>Forms</dt>
