--- app/services/cocina_generators/description/note.rb
@@ -1,31 +1,0 @@
-# frozen_string_literal: true
-
-module CocinaGenerators
-  class Description
-    # Generates the Cocina parameters for a note
-    class Note
-      def self.call(...)
-        new(...).call
-      end
-
-      def initialize(type:, value:, label: nil)
-        @type = type
-        @value = value
-        @label = label
-      end
-
-      def call
-        {
-          type:,
-          value:
-        }.tap do |note|
-          note[:displayLabel] = label if label
-        end
-      end
-
-      private
-
-      attr_reader :type, :value, :label
-    end
-  end
-end
