--- app/services/cocina_generators/description/person_contributor.rb
@@ -1,76 +1,0 @@
-# frozen_string_literal: true
-
-module CocinaGenerators
-  class Description
-    # Generates the Cocina contributor parameters for a person
-    class PersonContributor
-      def self.call(...)
-        new(...).call
-      end
-
-      # @param surname [String] the surname of the person
-      # @param forename [String] the forename of the person
-      # @param role [String] the role of the person from ROLES
-      # @param primary [Boolean] whether this is the first contributor
-      # @param orcid [String] the ORCID of the person
-      # @param cited [Boolean] whether the person should be cited
-      def initialize(surname:, forename:, role:, primary: false, orcid: nil, cited: true) # rubocop:disable Metrics/ParameterLists
-        @surname = surname
-        @forename = forename
-        @role = role
-        @primary = primary
-        @orcid = orcid
-        @cited = cited
-      end
-
-      def call
-        {
-          name: [
-            {
-              structuredValue: [
-                { value: forename, type: 'forename' },
-                { value: surname, type: 'surname' }
-              ]
-            }
-          ],
-          type: 'person',
-          role: role_params,
-          identifier: identifier_params,
-          status: ('primary' if primary),
-          note: note_params
-
-          # NOTE: affiliations.map { |affiliation_attrs| affiliation(**affiliation_attrs) }.presence
-        }.compact
-      end
-
-      private
-
-      attr_reader :forename, :surname, :role, :primary, :orcid, :cited
-
-      def role_params
-        cocina_role = ContributorRole.call(role:)
-        return if cocina_role.nil?
-
-        [cocina_role]
-      end
-
-      def identifier_params
-        return if orcid.blank?
-
-        [
-          {
-            value: orcid,
-            type: 'ORCID',
-            source: { uri: Settings.orcid.url }
-          }
-        ]
-      end
-
-      def note_params
-        return if cited
-
-        [{ type: 'citation status', value: 'false' }]
-      end
-    end
-  end
-end
