+++ spec/cocina_builders/organization_contributor_cocina_builder_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe CocinaGenerators::Description::OrganizationContributor do
+RSpec.describe OrganizationContributorCocinaBuilder do
   context 'with conference role' do
     subject(:cocina_params) { described_class.call(name: 'RailsConf', role: 'conference') }
