+++ spec/cocina_builders/person_contributor_cocina_builder_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe CocinaGenerators::Description::PersonContributor do
+RSpec.describe PersonContributorCocinaBuilder do
   subject(:contributor_params) do
     described_class.call(forename: 'Leland', surname: 'Stanford', role: 'funder', orcid:, cited:)
   end
